# perf(core): reduce memory allocations