--- a/lib/core/utils/query-selector-all-filter.js
+++ b/lib/core/utils/query-selector-all-filter.js
@@ -1,15 +1,32 @@
 import { matchesExpression, convertSelector } from './matches';
 
-function createLocalVariables(vNodes, anyLevel, thisLevel, parentShadowId) {
-  const retVal = {
-    vNodes: vNodes,
-    vNodesIndex: 0,
-    anyLevel: anyLevel,
-    thisLevel: thisLevel,
-    parentShadowId: parentShadowId
-  };
+function createLocalVariables(
+  vNodes,
+  anyLevel,
+  thisLevel,
+  parentShadowId,
+  recycledLocalVariable
+) {
+  const retVal = recycledLocalVariable || {};
+
+  retVal.vNodes = vNodes;
+  retVal.vNodesIndex = 0;
+  retVal.anyLevel = anyLevel;
+  retVal.thisLevel = thisLevel;
+  retVal.parentShadowId = parentShadowId;
+
   return retVal;
 }
+
+/**
+ * Allocating new objects in createLocalVariables is quite expensive given
+ * that matchExpressions is in the hot path.
+ *
+ * Keep track of previously allocated objects to avoid useless allocations
+ * and garbage collection. This is intentionally shared between calls of
+ * matchExpressions.
+ */
+const recycledLocalVariables = [];
 
 function matchExpressions(domTree, expressions, filter) {
   const stack = [];
@@ -18,7 +35,8 @@ function matchExpressions(domTree, expressions, filter) {
     vNodes,
     expressions,
     [],
-    domTree[0].shadowId
+    domTree[0].shadowId,
+    recycledLocalVariables.pop()
   );
   const result = [];
 
@@ -75,7 +93,8 @@ function matchExpressions(domTree, expressions, filter) {
         vNode.children,
         childAny,
         childOnly,
-        vNode.shadowId
+        vNode.shadowId,
+        recycledLocalVariables.pop()
       );
     }
     // check for "return"
@@ -83,6 +102,7 @@ function matchExpressions(domTree, expressions, filter) {
       currentLevel.vNodesIndex === currentLevel.vNodes.length &&
       stack.length
     ) {
+      recycledLocalVariables.push(currentLevel);
       currentLevel = stack.pop();
     }
   }