--- a/lib/core/utils/query-selector-all-filter.js
+++ b/lib/core/utils/query-selector-all-filter.js
@@ -2,12 +2,12 @@ import { matchesExpression, convertSelector } from './matches';
 
 function createLocalVariables(vNodes, anyLevel, thisLevel, parentShadowId) {
   const retVal = {
-    vNodes: vNodes.slice(),
+    vNodes: vNodes,
+    vNodesIndex: 0,
     anyLevel: anyLevel,
     thisLevel: thisLevel,
     parentShadowId: parentShadowId
   };
-  retVal.vNodes.reverse();
   return retVal;
 }
 
@@ -22,8 +22,8 @@ function matchExpressions(domTree, expressions, filter) {
   );
   const result = [];
 
-  while (currentLevel.vNodes.length) {
-    const vNode = currentLevel.vNodes.pop();
+  while (currentLevel.vNodesIndex < currentLevel.vNodes.length) {
+    const vNode = currentLevel.vNodes[currentLevel.vNodesIndex++];
     const childOnly = []; // we will add hierarchical '>' selectors here
     const childAny = [];
     const combinedLength =
@@ -79,7 +79,10 @@ function matchExpressions(domTree, expressions, filter) {
       );
     }
     // check for "return"
-    while (!currentLevel.vNodes.length && stack.length) {
+    while (
+      currentLevel.vNodesIndex === currentLevel.vNodes.length &&
+      stack.length
+    ) {
       currentLevel = stack.pop();
     }
   }