Profiling with [Firefox Profiler](https://profiler.firefox.com/) revealed some memory allocations in the hot path that can be avoided (see individual commits for more details). Removing these allocations results in lower memory usage (don't have hard numbers on this) and reduces the running time (by ~18% for certain use-cases):

```shell
# Before
hyperfine --warmup 1 'axe --browser firefox https://www.theguardian.com/international'                                   
Benchmark #1: axe --browser firefox https://www.theguardian.com/international
  Time (mean ± σ):     20.343 s ±  0.425 s    [User: 23.073 s, System: 3.503 s]
  Range (min … max):   19.959 s … 21.289 s    10 runs

# After
hyperfine --warmup 1 'axe --browser firefox https://www.theguardian.com/international'
Benchmark #1: axe --browser firefox https://www.theguardian.com/international
  Time (mean ± σ):     16.587 s ±  0.191 s    [User: 19.432 s, System: 3.307 s]
  Range (min … max):   16.365 s … 16.952 s    10 runs
```