@straker 2021-06-23
I'm a bit concerned with the added complexity of the latest 2 commits. Can you quantify the memory savings of them? Adding what is essentially an object pool just to save 80ms of run time may not be worth it unless the memory savings is huge. Same with the last commit to not allocate the arrays.