@sorin-davidoi 2021-06-24
> Can you quantify the memory savings of them?

Sure, all of the below measurements are taken on https://www.theguardian.com/international (~3700 DOM nodes).

The last commit (initializing `childOnly` and `childAny` lazily) reduces the memory allocated by `matchExpressions` by ~50% (from [~1400KB](https://share.firefox.dev/3qoJI5O) to [~700KB](https://share.firefox.dev/3gYP96Y)). This translates in a reduction of the running time of the function by a similar factor (switch the data source to "Timings" in the above links). I guess that's because very few of the iteration loops use both variables.

The other one (recyling local variables) reduces the memory allocated by `createLocalVariables` by ~99% (from [~215Kb](https://share.firefox.dev/3qoJI5O) to [under 1Kb](https://share.firefox.dev/3zXn34W)). This translates in a reduction of the running time of the fucntion by a similar factor - it spent the vast majority of its running time allocating memory that ended up being garbage collected almost immediately.

That being said, I can take them out and submit them separately if they are blocking the merging of the other three. There aren't a lot of cases where these kind of optimizations are worth it, but I think that here the impact they have on the running time and memory consumption justifies them in this case.