@straker 2021-06-28
@sorin-davidoi so for some reason this pr isn't showing up in our circle ci builds and can't run tests. Could you push an empty commit and see if that triggers the build properly?

```
git commit -m 'Trigger build' --allow-empty
```

If that doesn't work, could you enable "Allow edits from maintainers" on your fork and see if me pushing an empty commit changes anything.

Otherwise I'm not sure what else we can try other than reopening this pr with all your changes in a single commit to see if that works.