@sorin-davidoi 2021-06-28
The empty commits doesn't seem to have worked. "Allow edits from maintainers" is not available (presumably due to https://github.com/isaacs/github/issues/1681).