@sorin-davidoi 2021-06-29
Opened and closed https://github.com/dequelabs/axe-core/pull/3052  as it has the same issue. `npm run eslint && npm run build && npm run test && npm run test:integration && npm run test:node` runs successfully for me locally for what it's worth (`npm run test:examples` fails due to some `npm`-related reason).