@straker 2021-06-29
Alright. I pulled down your code to a new branch in this repo and pushed https://github.com/dequelabs/axe-core/pull/3053. That seems to be working and now CircleCI is picking it up and running tests. So I'm going to close this pr and we'll get the other one merged in.